# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your API Title",
    "version": "1.0.0",
    "description": "Your API description"
  },
  "paths": {
...
    "/users/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "surname",
                  "country",
                  "stateOfOrigin",
                  "employmentStatus",
                  "email",
                  "telephone",
                  "role",
                  "password"
                ],
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "otherName": {
                    "type": "string"
                  },
                  "surname": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "stateOfOrigin": {
                    "type": "string"
                  },
                  "LGA": {
                    "type": "string",
                    "default": ""
                  },
                  "maritalStatus": {
                    "type": "string",
                    "enum": [
                      "single",
                      "married",
                      "divorced"
                    ],
                    "default": "single"
                  },
                  "employmentStatus": {
                    "type": "string",
                    "enum": [
                      "employed",
                      "unemployed"
                    ],
                    "default": "unemployed"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email address of the user"
                  },
                  "telephone": {
                    "type": "string"
                  },
                  "profilePicture": {
                    "type": "string",
                    "format": "binary"
                  },
                  "role": {
                    "type": "string",
                    "default": "User"
                  },
                  "password": {
                    "type": "string"
                  },
                  "accountStatus": {
                    "type": "string",
                    "description": "Status of the user's account",
                    "default": "Active"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "User already exists"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/users/profile/{userId}": {
      "put": {
        "summary": "Update user profile",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "default": ""
                  },
                  "otherName": {
                    "type": "string",
                    "default": ""
                  },
                  "surname": {
                    "type": "string",
                    "default": ""
                  },
                  "country": {
                    "type": "string",
                    "default": ""
                  },
                  "stateOfOrigin": {
                    "type": "string",
                    "default": ""
                  },
                  "LGA": {
                    "type": "string",
                    "default": ""
                  },
                  "maritalStatus": {
                    "type": "string",
                    "enum": [
                      "single",
                      "married",
                      "divorced"
                    ],
                    "default": "single"
                  },
                  "employmentStatus": {
                    "type": "string",
                    "enum": [
                      "employed",
                      "unemployed"
                    ],
                    "default": "unemployed"
                  },
                  "email": {
                    "type": "string",
                    "default": ""
                  },
                  "telephone": {
                    "type": "string",
                    "default": ""
                  },
                  "profilePicture": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
...
}